Move burn transaction modal to component
