+++ v2/components/Burn/Burn.stories.tsx
@@ -16,16 +16,11 @@ Primary.args = {
   susdBalance: 2000,
   debtBalance: 100,
   isLoading: false,
-  txnStatus: 'unsent',
   onSubmit: () => {},
   onBurnAmountSusdChange: () => {},
   burnAmountSusd: '',
   snxUnstakingAmount: '',
   transactionFee: wei(0),
-  modalOpen: false,
-  error: null,
-  gasError: null,
-  settle: () => {},
   isGasEnabledAndNotFetched: true,
   onBadgeClick: () => {},
   stakedSnx: 100,
